requirements.txt
Django
django-filter
djangorestframework
Pillow
easy-thumbnails
gunicorn
django-environ